remove not needed header file, rename to home
